タイトル画面に文字追加

+++ Game/Map.h
@@ -32,6 +32,6 @@ private:
 	bool m_bScrollChangeIs;		//スクロールチェンジ判定
 	float m_fMovePos_x;			//スクロールした移動量保存用変数
 	float m_fMovePos_y;
-	float m_fMoveAddPos_x;		//2回目以降生成位置調整用変数
-	float m_fMoveAddPos_y;
+	int m_fMoveAddPos_x;		//2回目以降生成位置調整用変数
+	int m_fMoveAddPos_y;
 };
